% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_gini.R
\name{if_gini}
\alias{if_gini}
\title{Influence Function for the Gini Coefficient}
\usage{
if_gini(y, weights = NULL, na.rm = TRUE)
}
\arguments{
\item{y}{Numeric vector of income or variable of interest.}

\item{weights}{Numeric vector of sampling weights. If \code{NULL} (default),
equal weights are assumed (simple random sampling).}

\item{na.rm}{Logical. Should missing values be removed? Default is \code{TRUE}.}
}
\value{
A numeric vector of the same length as \code{y} containing the
influence function values for each observation.
}
\description{
Computes the influence function for the Gini coefficient, useful for
variance estimation in complex survey designs. The function implements
the linearization approach described in \insertCite{langel2013variance;textual}{inequantiles}.
}
\details{
The influence function for the Gini coefficient is computed using the
linearization method, follwing \insertCite{deville1999variance;textual}{inequantiles}
framework and as defined by \insertCite{langel2013variance;textual}{inequantiles}.
For observation \eqn{k} with value \eqn{y_k} and
weight \eqn{w_k}, the influence function is:

\deqn{z_k = \frac{2W_k(y_k - \bar{Y}_k) + \hat{Y} - \hat{N}y_k - G(\hat{Y} + y_k\hat{N})}{\hat{N}\hat{Y}}}

where:
\itemize{
\item \eqn{W_k = \sum_{i=1}^k w_i} is the cumulative sum of weights up to rank \eqn{k}
\item \eqn{\bar{Y}_k =\frac{\sum_{l \in S} w_l y_l 1\left(W_l \leqslant W_j\right)}{W_k}} is the weighted mean of values up to rank \eqn{k}
\item \eqn{\hat{N} = \sum_i w_i} is the total sum of weights
\item \eqn{\hat{Y} = \sum_i w_i y_i} is the weighted total of the variable
\item \eqn{G} is the Gini coefficient estimate
}

The observations are ranked by their values before computing the influence function.
}
\examples{

data(synthouse)

eq <- synthouse$eq_income ### Income data

# Simple example
z <- if_gini(eq)

# With weights
w <- synthouse$weight
z_weighted <- if_gini(y = eq, weights = w)



}
\references{
\insertRef{deville1999variance}{inequantiles}

\insertRef{langel2013variance}{inequantiles}
}
\seealso{
\code{\link{if_qsr}} for the quintile share ratio influence function
and \code{\link{if_qri}} for the quantile ratio index influence function.
}
