% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qri_infleunce_function.R
\name{if_qri}
\alias{if_qri}
\title{Influence Function for the Quantile Ratio Index}
\usage{
if_qri(y, weights = NULL, type = 6)
}
\arguments{
\item{y}{A numeric vector of data values}

\item{weights}{A numeric vector of sampling weights (optional)}

\item{type}{Quantile estimation type: integer 4–9 or "HD" for Harrell–Davis (default: 7)}

\item{k}{Optional integer, index of the unit for which to compute the influence function.
If NULL (default), the function returns the IF for all observations.}
}
\value{
A numeric value (if k is provided) or a numeric vector (if k is NULL) of influence function values
}
\description{
Computes the influence function of the QRI for all observations.
}
\details{
The influence function is computed as:
\deqn{I(QRI)_i = - \int_0^1 \frac{IF(Q(p/2))_i Q(1-p/2) - IF(Q(1-p/2))_i Q(p/2)}{Q(1-p/2)^2} dp}

The density is estimated using a Gaussian kernel with bandwidth
\eqn{h = 0.79 \cdot IQR \cdot \hat{N}^{-1/5}}.

#' #' @references
Scarpa, S., Ferrante, M.R., & Sperlich, S. (2025). Inference for the Quantile Ratio
Inequality Index in the Context of Survey Data. \emph{Journal of Survey Statistics and Methodology}
}
