% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_qsr.R
\name{if_qsr}
\alias{if_qsr}
\title{Influence Function for the Quintile Share Ratio (QSR)}
\usage{
if_qsr(y, weights = NULL, type = 4)
}
\arguments{
\item{y}{A numeric vector of income or other continuous variable.}

\item{weights}{A numeric vector of sampling weights. If \code{NULL} (default),
equal weights are assumed for all observations.}

\item{type}{Quantile estimation type: integer 4–9 or "HD" for Harrell–Davis (default: 4)}
}
\value{
A numeric vector of the same length as \code{y} containing the
linearized variable \eqn{\widehat{z}_k} for each observation.
}
\description{
Computes the linearized variable (influence function) for the quintile share
ratio (QSR) using the linearization approach based on
\insertCite{deville1999variance;textual}{inequantiles}
and the derivation in \insertCite{langel2011quintile;textual}{inequantiles}
}
\details{
According to the definition by \insertCite{langel2011quintile;textual}{inequantiles} ,
the influence function of QSR is given by:

\deqn{I(\mathrm{QSR})_k=\frac{y_k-I\left(Y_{0.8}\right)_k}{Y_{0.2}}-\frac{\left(Y-Y_{0.8}\right) I\left(Y_{0.2}\right)_k}{Y_{0.2}^2}}

where \eqn{I(Y_p)_k = p Q(p) - (Q(p) - y_k) \mathbb{1}[y_k \leq Q(p)]} is the
influence function of the partial total up to quantile \eqn{p} and
\eqn{Y_{\alpha} = \sum_{j} y_j\mathbb{1}[y_k \leq {Q}(\alpha)]}.

The estimated linearized variable is:
\deqn{\widehat{z}_k=\frac{y_k-\left\{0.8 \widehat{Q}(0.8)-\left(\widehat{Q}(0.8)-y_k\right) \mathbb{1}\left[y_k \leq \widehat{Q}(0.8)\right]\right\}}{\widehat{Y}_{0.2}}-\frac{\left(\widehat{Y}-\widehat{Y}_{0.8}\right)\left\{0.2 \widehat{Q}(0.2)-\left(\widehat{Q}(0.2)-y_k\right) \mathbb{1}\left[y_k \leq \widehat{Q}(0.2)\right]\right\}}{\widehat{Y}_{0.2}^2}}

where \eqn{\hat{Y}_{\alpha} = \sum_{j \in s} w_j y_j\mathbb{1}[y_k \leq \widehat{Q}(\alpha)]} and
\eqn{\widehat{Q}(p)} is the weighted p-th quantile estimator, computed using the internal function \code{csquantile()}.
}
\examples{

data(synthouse)

eq <- synthouse$eq_income ### Income data

# Simple example
z <- if_qsr(eq)

# With weights
w <- synthouse$weight
z_weighted <- if_qsr(y = eq, weights = w)

}
\references{
\insertRef{deville1999variance}{inequantiles}

\insertRef{langel2011quintile}{inequantiles}
}
\seealso{
\code{\link{qsr}} for the QSR estimator, \code{\link{csquantile}}
for weighted quantile estimation.
}
