% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palma_ratio.R
\name{palma_ratio}
\alias{palma_ratio}
\title{Palma Ratio Estimator}
\usage{
palma_ratio(y, weights = NULL, type = 6, na.rm = TRUE)
}
\arguments{
\item{y}{A numeric vector of data values}

\item{weights}{A numeric vector of sampling weights (optional)}

\item{type}{Quantile estimation type: integer 4-9 or HD for Harrell-Davis (default: 6).
See \code{csquantile} documentation for a complete description.}

\item{na.rm}{Logical, should missing values be removed? (default: TRUE)}
}
\value{
A scalar numeric value representing the estimated inequality by the
Palma ratio.
}
\description{
Computes the Palma ratio estimator for measuring inequality
on simple and complex sampling data
}
\details{
Consider a random sample \eqn{s} of size \eqn{n}, and let \eqn{w_j}, \eqn{j \in s}, define
the sampling weight and \eqn{y_j} be the observed characteristics (i.e. income)
associated to the \eqn{j}-th individual, \eqn{j = 1, \ldots, n}.
According to \insertCite{cobham2013all;textual}{inequantiles} definition,
the Palma index divides the share earned by the richest 10\\% by the share of the
poorest 40\\%. Its estimator is defined as:

\deqn{\widehat{Palma} =
\frac{\sum_{j \in s}w_j y_j \mathds{1}\left\{ y_j \geq \widehat{Q}(0.9)\right\} }{\sum_{j \in s} w_j y_j\mathds{1}\left\{ y_j \leq \widehat{Q}(0.4)\right\} }}

where the estimated quantiles \eqn{\widehat{Q}(p)} are computed via the function
\code{csquantile()}, which accounts for sampling weights and the specified
quantile type. This allows \eqn{\widehat{Palma}} to be used both for simple
random samples and for complex survey data with design weights.

See  \insertCite{palma2006globalizing;textual}{inequantiles} and \insertCite{palma2011homogeneous;textual}{inequantiles}
for an introduction to the Palma ratio.
}
\examples{

data(synthouse)

eq <- synthouse$eq_income ### Income data

# Compute unweighted Palma index with default type 6 quantile estimator
palma_ratio(y = eq)

# Consider the sampling weights and change quantile estimation type
w <- synthouse$weight
palma_ratio(y = eq, weights = w, type = 5)

# Compare the Palma index across macro-regions (NUTS1)
tapply(1:nrow(synthouse), synthouse$NUTS1, function(area) {
  palma_ratio(y = synthouse$eq_income[area],
      weights = synthouse$weight[area],
      type = 6)
})

}
\references{
\insertRef{palma2006globalizing}{inequantiles}

\insertRef{palma2011homogeneous}{inequantiles}

\insertRef{cobham2013all}{inequantiles}
}
