% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_quantiles.R
\name{ratio_quantiles}
\alias{ratio_quantiles}
\title{Ratio of Quantiles (e.g., P90/P10)}
\usage{
ratio_quantiles(
  y,
  weights = NULL,
  prob_numerator = 0.9,
  prob_denominator = 0.1,
  type = 6,
  na.rm = TRUE
)
}
\arguments{
\item{y}{A numeric vector of data values}

\item{weights}{A numeric vector of sampling weights (optional). If \code{NULL},
all observations are equally weighted.}

\item{prob_numerator}{The percentile to be considered at the numerator (default \code{0.90})}

\item{prob_denominator}{The percentile to be considered at the denominator (default \code{0.10})}

\item{type}{Quantile estimation type: integer \code{4}--\code{9} or
\code{"HD"} for Harrell--Davis (default: \code{6}). See \code{\link{csquantile}}.}

\item{na.rm}{Logical, should missing values be removed? (default: TRUE)}
}
\value{
A scalar numeric value representing the estimated ratio of quantiles
}
\description{
Estimates ratio of quantiles (e.g., P90/P10) on simple and complex sampling data
}
\details{
Consider a random sample \eqn{s} of size \eqn{n}, and let \eqn{w_j}, \eqn{j \in s}, define
the sampling weight and \eqn{y_j} be the observed characteristics (i.e. income)
associated to the \eqn{j}-th individual, \eqn{j = 1, \ldots, n}.
Let \eqn{p_{n}} be the order of the quantile at the numerator and
\eqn{p_{d}} be the order of the quantile at the denominator. For example, set \eqn{p_{n} = 0.90} and
\eqn{p_{d} = 0.10}. Then the popular P90/P10 ratio can be estimated by

\deqn{\widehat{{P}90/{P}10} = \frac{\widehat{Q}(p_n=0.9)}{\widehat{Q}(p_d=0.1)}  }

where the estimated quantiles \eqn{\widehat{Q}(p)} are computed via the function
\code{csquantile()}, which accounts for sampling weights and the specified
quantile type.
}
\examples{

data(synthouse)

eq <- synthouse$eq_income # Income data

# Compute unweighted P90/P10 with default type 6 quantile estimator
ratio_quantiles(y = eq)

# Consider the sampling weights, change quantile estimation type and orders of quantiles
w <- synthouse$weight
ratio_quantiles(y = eq, weights = w, prob_numerator = 0.6, prob_denominator = 0.1, type = 5)

# Compare the P90/P10 across macro-regions (NUTS1)
tapply(1:nrow(synthouse), synthouse$NUTS1, function(area) {
  ratio_quantiles(y = synthouse$eq_income[area],
      weights = synthouse$weight[area])
})


}
\seealso{
Other inequality indicators based on quantiles: 
\code{\link{inequantiles}()},
\code{\link{plot_inequality_curve}()},
\code{\link{qri}()},
\code{\link{share_ratio}()},
\code{\link{superpop_qri}()}
}
\concept{inequality indicators based on quantiles}
