% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaled_bootstrap.R
\name{rescaled_bootstrap}
\alias{rescaled_bootstrap}
\title{Rescaled Bootstrap Variance Estimation}
\usage{
rescaled_bootstrap(
  data,
  y,
  strata,
  N_h = NULL,
  psu = NULL,
  weights = NULL,
  estimator,
  by_strata = TRUE,
  B = 200,
  m_h = NULL,
  seed = NULL
)
}
\arguments{
\item{data}{A data frame containing the survey data}

\item{strata}{A character string or vector specifying the stratification variable}

\item{N_h}{Optional vector of stratum population size, needed for computing the  finite population correction (for stratified srs).
Can be a vector of length H (one per stratum) or a single value for all strata.
Default is NULL (no FPC applied)}

\item{psu}{A character string specifying the Primary Sampling Units variable (required for a multistage sampling design)}

\item{weights}{A character string specifying the weight variable (required if design = "complex")}

\item{estimator}{A function that takes the data (and weights if needed)
and returns the estimate. For simple designs: function(data).
For complex designs: function(data, weights)}

\item{B}{Integer, number of bootstrap replicates (default = 200)}

\item{m_h}{Optional vector of bootstrap sample sizes per stratum. If NULL, computed as floor((n_h-2)^2/(n_h-1))}

\item{seed}{Optional integer for reproducibility}
}
\value{
A list containing:
\item{variance}{Bootstrap variance estimate}
\item{boot_estimates}{Vector of B bootstrap estimates}
\item{B}{Number of bootstrap replicates}
\item{by_strata}{if variance is computed by stratum or not}
\item{design}{if the sampling design is complex (with sampling weights) or simple}
\item{strata_info}{Returns information about number of observations/PSUs per stratum}
\item{call}

#'#' @references
Rao, J.N.K. & Wu, C.F.J. (1988). Resampling inference with complex survey data.
JASA, 83(401), 231-241.
Rao, J.N.K., Wu, C.F.J. & Yue, K. (1992). Some recent work on resampling
methods for complex surveys. Survey Methodology, 18(2), 209-217.
Kolenikov, S. (2010). Resampling variance estimation for complex survey data.
The Stata Journal, 10(2), 165-199.
Scarpa, Silvia, Maria Rosaria Ferrante, and Stefan Sperlich.
"INFERENCE FOR THE QUANTILE RATIO INEQUALITY INDEX IN THE CONTEXT OF SURVEY DATA.",
Journal of Survey Statistics and Methodology (2025): smaf024.
}
\description{
Implements the rescaled bootstrap method for variance estimation in survey data,
supporting both stratified simple random sampling and two-stage complex designs.
Based on Rao & Wu (1988, 1992).
}
