% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_estimator.R
\name{csquantile}
\alias{csquantile}
\title{Estimator of quantile in case of simple and complex sampling design}
\usage{
csquantile(y, weights = NULL, probs = seq(0, 1, 0.1), type = 5, na.rm = FALSE)
}
\arguments{
\item{y}{Numeric vector of observations.}

\item{weights}{Optional numeric vector of sampling weights (default: NULL for equal weights)}

\item{probs}{Numeric vector of probabilities (default = seq(0, 1, 0.1)).}

\item{type}{Quantile algorithm: integer 4-9 or "HD" for Harrell-Davis (default: 5)}

\item{na.rm}{Logical indicating whether to remove NA values (default: FALSE)}
}
\value{
A named numeric vector of quantiles corresponding to \code{probs}.
}
\description{
Computes quantiles for weighted or unweighted data, allowing several interpolation, as studied by
Scarpa et al (2025). Supports standard quantile types (4-9) and the
Harrell-Davis estimator ("HD") when no weights are considered.
}
\references{
Hyndman, R.J. & Fan, Y. (1996). Sample quantiles in statistical packages.
\emph{The American Statistician}, 50(4), 361–365.

Scarpa, S., Ferrante, M.R., & Sperlich, S. (2025). Inference for the Quantile Ratio
Inequality Index in the Context of Survey Data. \emph{Journal of Survey Statistics and Methodology}.
}
