% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_qri.R
\name{if_qri}
\alias{if_qri}
\title{Influence Function for the Quantile Ratio Index}
\usage{
if_qri(y, weights = NULL, type = 6)
}
\arguments{
\item{y}{A numeric vector of data values}

\item{weights}{A numeric vector of sampling weights (optional). If \code{NULL}, equal weights are assumed.}

\item{type}{Quantile estimation type: integer 4–9 or HD for Harrell–Davis (default: 6)
See \code{csquantile} documentation for a complete description.}
}
\value{
A numeric vector of influence function values (one for each observation)
}
\description{
Computes the influence function of the quantile ratio index (QRI) in the context
of finite population for all observations,
as defined in \insertCite{scarpa2025inference;textual}{inequantiles}, under simple and complex sampling.
See \insertCite{deville1999variance;textual}{inequantiles} for an introduction to the definition of influence
function in finite population theory.
}
\details{
The influence function for the QRI is computed on each observation as
\deqn{
  \widehat{z}_i = - \int_0^1
  \frac{
  \left(
    \frac{\frac{p}{2} - \mathbf{1}(y_i \leq \widehat{Q}(p/2))}
    {\widehat{f}(\widehat{Q}(p/2)) \, \widehat{N}}
  \right)
  \widehat{Q}(1 - p/2)
  -
  \left(
    \frac{(1 - \frac{p}{2}) - \mathbf{1}(y_i \leq \widehat{Q}(1 - p/2))}
    {\widehat{f}(\widehat{Q}(1 - p/2)) \, \widehat{N}}
  \right)
  \widehat{Q}(p/2)
  }{
    \widehat{Q}(1 - p/2)^2
  } \, dp
}

where:
\itemize{
\item \eqn{\widehat{Q}(p)} is the weighted sample quantile of order \eqn{p},
computed using the internal function \code{csquantile()},
\item \eqn{\widehat{f}(\cdot)} denotes the estimated income density function,
\item \eqn{\widehat{N} = \sum_i w_i} is the estimated population size.
}

The density function \eqn{\widehat{f}(y)} is estimated via a Gaussian kernel smoother:
\deqn{
  \widehat{f}(y) =
  \frac{1}{\widehat{N}} \sum_{j \in s} w_j
  K\!\left(\frac{y - y_j}{h}\right)
  =
  \frac{1}{\widehat{N}\, h \sqrt{2\pi}}
  \sum_{j \in s} w_j
  \exp\!\left\{ -\frac{(y - y_j)^2}{2h^2} \right\},
}
where \eqn{K(\cdot)} is the Gaussian kernel.

The bandwidth is chosen as:
\deqn{
  h = 0.79 \cdot \mathrm{IQR} \cdot \widehat{N}^{-1/5},
}
where \eqn{\mathrm{IQR}} is the interquartile range of the weighted sample.
}
\examples{

# On synthetic data
eq_synth <- rlnorm(30, 9, 0.7)
IF_synth <- if_qri(y = eq_synth)

# On real data
eq <- synthouse$eq_income[1:30] ## Take some observations (as example)
w <- synthouse$weight[1:30]
IF_qri <- if_qri(y = eq, weights = w, type = 6)



}
\references{
\insertRef{deville1999variance}{inequantiles}

\insertRef{scarpa2025inference}{inequantiles}
}
