% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_infleunce_function.R
\name{if_quantile}
\alias{if_quantile}
\title{Influence Function for Quantiles}
\usage{
if_quantile(y, weights = NULL, probs, type = 6, na.rm = TRUE)
}
\arguments{
\item{y}{A numeric vector of data values}

\item{weights}{A numeric vector of sampling weights (optional)}

\item{probs}{A numeric value specifying the probability for the quantile (e.g., 0.5 for median)}

\item{type}{Quantile estimation type: integer 4-9 or "HD" for Harrell-Davis (default: 6)}

\item{na.rm}{Logical, should missing values be removed? (default: TRUE)}
}
\value{
A numeric vector containing the estimtaed quantile influence function values for each observation.
}
\description{
Computes the influence function of sample quantiles, allowing for both
simple random sampling and complex survey designs with sampling weights, in
the context of finite population  (see Deville, 1999, for a full explanation).
}
\details{
From the definiton in  Van der Vaart (2009), the population influence function of the quantile \eqn{Q(p)} is defined as:

\deqn{IF(Q(p))_i = \frac{p - \mathbf{1}(y_i \leq Q(p))}{f(Q(p)) \, N},}

where \eqn{f(Q(p))} is the population density function evaluated at the quantile and
\eqn{N} is the population size.

In the sample, this is estimated as:

\deqn{\widehat{IF}(Q(p))_i = \frac{p - \mathbf{1}(y_i \leq \widehat{Q}(p))}{\widehat{f}(\widehat{Q}(p)) \, \widehat{N}},}

where \eqn{\widehat{Q}(p)} is the weighted sample quantile estimated by
\code{csquantile()}, and \eqn{\widehat{N} = \sum_{i \in s} w_i} is the estimated population size.

The density \eqn{\widehat{f}(y)} is estimated using a Gaussian kernel density function:

\deqn{
\widehat{f}(y) = \frac{1}{\widehat{N} \, h \sqrt{2\pi}}
\sum_{j \in s} w_j \exp\!\left\{ -\frac{(y - y_j)^2}{2h^2} \right\},
}

with bandwidth \eqn{h = 0.79 \cdot IQR \cdot \widehat{N}^{-1/5}}
}
\examples{

# On synthetic data
eq_synth <- rlnorm(30, 9, 0.7)
IF_synth <- if_quantile(y = eq_synth, probs = 0.3)

# On real data
data(synthouse)
eq <- synthouse$eq_income[1:30] ## Take some observations (as example)
w <- synthouse$weight[1:30]
IF_quantile <- if_quantile(y = eq, weights = w, type = 6, probs = 0.5)



}
\references{
Deville, J.C., (1999), “Variance estimation for complex statistics and estimators:
Linearization and residual techniques”, \emph{Survey methodology}, 25, 193–204

Van der Vaart, A. W., (2000),"Asymptotic statistics", (Vol. 3),
Cambridge University Press

Osier, G., (2009), “Variance estimation for complex indicators of poverty and inequality using
linearization techniques”, \emph{Survey Research Methods}, 3, 167–195
}
